requirements.txt
os
python-dotenv
requests
bs4